% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{ASEGeneCounts}
\alias{ASEGeneCounts}
\title{Allele-specific expression (ASE) gene counts}
\usage{
ASEGeneCounts()
}
\value{
A numeric matrix
}
\description{
Allele-specific expression (ASE) gene counts
}
\examples{
ase_counts <- ASEGeneCounts()
head(ase_counts)
}
