% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{clear_osf_cache}
\alias{clear_osf_cache}
\title{Clear OSF cache}
\usage{
clear_osf_cache(
  bfc = BiocFileCache::BiocFileCache(tools::R_user_dir("HumanRetinaLRSData", which =
    "cache"), ask = FALSE)
)
}
\arguments{
\item{bfc}{BiocFileCache object}
}
\value{
NULL (invisibly). Called for side effect of clearing cache.
}
\description{
Clear OSF cache
}
\examples{
clear_osf_cache()
}
