% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{FTRGCGeneLevelData}
\alias{FTRGCGeneLevelData}
\title{Gene-level counts for flow through (FT) and retinal ganglion cell (RGC) samples}
\usage{
FTRGCGeneLevelData()
}
\value{
A SummarizedExperiment object
}
\description{
Gene-level counts for flow through (FT) and retinal ganglion cell (RGC) samples
}
\examples{
ftrgc_data <- FTRGCGeneLevelData()
ftrgc_data
}
