% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{FTRGCIsoformLevelData}
\alias{FTRGCIsoformLevelData}
\title{Isoform-level counts for flow through (FT) and retinal ganglion cell (RGC) samples}
\usage{
FTRGCIsoformLevelData()
}
\value{
A SummarizedExperiment object
}
\description{
Isoform-level counts for flow through (FT) and retinal ganglion cell (RGC) samples
}
