% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{list_osf_files}
\alias{list_osf_files}
\title{List available files on OSF}
\usage{
list_osf_files()
}
\value{
Character vector of file names
}
\description{
List available files on OSF
}
\examples{
list_osf_files()
}
