% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{ROGeneLevelData}
\alias{ROGeneLevelData}
\title{Gene-level counts for retinal organoid (RO) samples}
\usage{
ROGeneLevelData()
}
\value{
A SummarizedExperiment object
}
\description{
Gene-level counts for retinal organoid (RO) samples
}
\examples{
ro_data <- ROGeneLevelData()
ro_data
}
