% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{load_object}
\alias{load_object}
\title{Load object from OSF}
\usage{
load_object(osf_file_name)
}
\arguments{
\item{osf_file_name}{Name of the file on OSF}
}
\value{
The R object stored in the RDS file
}
\description{
Load object from OSF
}
\keyword{internal}
