% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{ROIsoformLevelData}
\alias{ROIsoformLevelData}
\title{Isoform-level counts for retinal organoid (RO) samples}
\usage{
ROIsoformLevelData()
}
\value{
A SummarizedExperiment object
}
\description{
Isoform-level counts for retinal organoid (RO) samples
}
